requirements.txt
aiogram==3.13.1
python-dotenv==1.0.0
aiofiles==23.2.1
python-magic==0.4.27
